+++ src/components/authorPage/AuthorCollection.js
@@ -1,6 +1,6 @@
 'use strict'
 
-import { AUTHOR_COLLECTION, LOADING_MORE_ARTICLES, LOAD_MORE_ARTICLES, REQUEST_PAGE_START_FROM } from '../../constants/author-page'
+import { AUTHOR_COLLECTION, LOADING_MORE_ARTICLES, LOAD_MORE_ARTICLES, NUMBER_OF_FIRST_RESPONSE_PAGE } from '../../constants/author-page'
 import { CHARACTERS_LIMIT } from '../../constants/index'
 
 import { Link } from 'react-router'
@@ -15,7 +15,7 @@ import { shortenString } from '../../lib/string-processor'
 import styles from './AuthorCollection.scss'
 
 const AuthorCollection = (props) => {
-  const { relateds, isFinish, isFetching, currentPage, handleLoadmore, totalResults } = props
+  const { relateds, hasMore, isFetching, currentPage, handleLoadmore, totalResults } = props
 
   const titleText = AUTHOR_COLLECTION +`（${totalResults}）`
   let listItems = relateds
@@ -26,15 +26,13 @@ const AuthorCollection = (props) => {
 
   // Page bottom display options
 
-  let loadmoreBtnDisplay = (currentPage <= REQUEST_PAGE_START_FROM && !isFinish &&!isFetching) ? true : false
-  let sensorDisplay = (currentPage > REQUEST_PAGE_START_FROM && !isFinish) ? true : false
+  let loadmoreBtnDisplay = (currentPage <= NUMBER_OF_FIRST_RESPONSE_PAGE && hasMore &&!isFetching) ? true : false
+  let sensorDisplay = (currentPage > NUMBER_OF_FIRST_RESPONSE_PAGE && hasMore) ? true : false
   let loaderDisply = isFetching ? true : false
-
-  // const loadmoreBtn = isFinish || isFetching || currentPage<-1 ? null : <div className={classNames(styles['load-more'], 'text-center')} onClick={handleLoadmore}>{LOAD_MORE_ARTICLES}</div>
 
   // Callback for sensor is triggered to seen
   let handleSeen = (isVisible) => {
-    if (currentPage>REQUEST_PAGE_START_FROM && isVisible === true) {
+    if (currentPage>NUMBER_OF_FIRST_RESPONSE_PAGE && isVisible === true) {
       return handleLoadmore()
     }
   }
@@ -88,30 +86,8 @@ AuthorCollection.propTypes = {
   currentPage: PropTypes.number,
   handleLoadmore: PropTypes.func,
   isFetching: PropTypes.bool,
-  isFinish: PropTypes.bool,
-  related: PropTypes.arrayOf(PropTypes.shape({
-    categories: PropTypes.array,
-    designers: PropTypes.array,
-    engineers: PropTypes.array,
-    extendByline: PropTypes.string,
-    heroImage: PropTypes.object,
-    hilightResult: PropTypes.object,
-    id: PropTypes.string,
-    isFeatured: PropTypes.bool,
-    objectID: PropTypes.string,
-    ogDescription: PropTypes.string,
-    ogTitle: PropTypes.string,
-    photographers: PropTypes.array,
-    publishDate: PropTypes.string,
-    related: PropTypes.array,
-    slug: PropTypes.string,
-    state: PropTypes.string,
-    style: PropTypes.string,
-    subtitle: PropTypes.string,
-    tags: PropTypes.array,
-    topics: PropTypes.array,
-    writters: PropTypes.array
-  })),
+  hasMore: PropTypes.bool,
+  related: PropTypes.array,
   totalResults: PropTypes.number
 }