+++ src/constants/author-page.js
@@ -1,8 +1,8 @@
 'use strict'
 
-export const REQUEST_PAGE_START_FROM = 0
+export const NUMBER_OF_FIRST_RESPONSE_PAGE = 0
 export const MAX_ARTICLES_PER_FETCH = 5
-export const RETURN_DELAY = 1000 //ms
+export const RETURN_DELAY_TIME = 1000 //ms
 
 export const LOAD_MORE_ARTICLES = '載入更多文章'
 export const LOADING_MORE_ARTICLES = '載入更多文章'