+++ src/containers/Author.js
@@ -4,21 +4,19 @@ import { AUTHOR_PAGE } from '../constants/page-types'
 import AuthorCollection from '../components/authorPage/AuthorCollection'
 import AuthorData from '../components/authorPage/AuthorData'
 import { LIGHT } from '../constants/page-themes'
-import React from 'react'
+import React, { PropTypes } from 'react'
 import Sponsor from '../components/Sponsor'
 import authorDefaultImg from '../../static/asset/author-default-img.svg'
 import classNames from 'classnames'
 import commonStyles from '../components/article/Common.scss'
 import { connect } from 'react-redux'
 import { denormalizeArticles } from '../utils/denormalize-articles'
-import { fetchAuthorCollectionIfNeeded } from '../actions/author'
+import { fetchAuthorCollectionIfNeeded } from '../actions/author-articles'
 import get from 'lodash/get'
-import omit from 'lodash/omit'
 import { setHeaderInfo } from '../actions/header'
 
 const _ = {
-  get,
-  omit
+  get
 }
 
 class Author extends React.Component {
@@ -28,9 +26,8 @@ class Author extends React.Component {
   }
   componentDidMount() {
     const authorId = this.props.params['authorId']
-    let { setHeaderInfo, authorPage, fetchAuthorCollectionIfNeeded } = this.props
-    const currentPage = _.get(authorPage, [ authorId, 'currentPage' ], -1)
-    if (currentPage < 0) {
+    let { setHeaderInfo, articlesByAuthor, fetchAuthorCollectionIfNeeded } = this.props
+    if (!articlesByAuthor[authorId]) {
       fetchAuthorCollectionIfNeeded(authorId)
     }
     setHeaderInfo({
@@ -40,17 +37,16 @@ class Author extends React.Component {
   }
   render() {
     const authorId = this.props.params['authorId']
-    const { entities, authorPage } = this.props
-    const isFetching = _.get(authorPage, 'isFetching', false)
-    let { isFinish, currentPage, collectIndexList, totalResults } = _.get(authorPage, authorId, {})
+    const { entities, articlesByAuthor } = this.props
+    let { hasMore, isFetching, currentPage, collectIndexList, totalResults } = _.get(articlesByAuthor, authorId, {})
     let authorCollection = denormalizeArticles(collectIndexList, entities)
-    const thisAuthorInEntities = _.get(entities, [ 'authors', authorId ], {})
+    const authorEntity = _.get(entities, [ 'authors', authorId ], {})
     const authorData = {
       authorId: authorId,
-      authorName: _.get(thisAuthorInEntities, 'name'),
-      authorImg: _.get(thisAuthorInEntities, 'imageUrl', authorDefaultImg),
-      authorMail: _.get(thisAuthorInEntities, 'email'),
-      authorBio: _.get(thisAuthorInEntities, 'bio.md')
+      authorName: _.get(authorEntity, 'name'),
+      authorImg: _.get(authorEntity, 'imageUrl', authorDefaultImg),
+      authorMail: _.get(authorEntity, 'email'),
+      authorBio: _.get(authorEntity, 'bio.md')
     }
     let handleLoadmore = () => {
       return this.props.fetchAuthorCollectionIfNeeded(authorId)
@@ -62,7 +58,7 @@ class Author extends React.Component {
         <AuthorCollection
           relateds={authorCollection}
           currentId={authorId}
-          isFinish={isFinish}
+          hasMore={hasMore}
           isFetching={isFetching}
           currentPage={currentPage}
           handleLoadmore={handleLoadmore}
@@ -74,10 +70,15 @@ class Author extends React.Component {
   }
 }
 
+Author.propTypes = {
+  entities: PropTypes.object.isRequired,
+  articlesByAuthor: PropTypes.object.isRequired
+}
+
 function mapStateToProps(state) {
   return {
     entities: _.get(state, 'entities', {}),
-    authorPage: _.get(state, 'author', {})
+    articlesByAuthor: _.get(state, 'articlesByAuthor', {})
   }
 }