+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [Update] Update api/api.js. Handle redis connection
 - [Update] Show categories on navbar when device is not mobile
 - [Add] Add unit test for api actions
+- [Update] Update and rename state authors to articlesByAuthor
 
 ### 1.1.4
 - [Update] Update src/utils/image-processor.js. Add handling for empty or wrong parameter.